Azure Functions/Proj_LLMSQL/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.15.0
jinja2==3.1.2
numpy==1.25.2
packaging==23.1
openai==0.27.8
langchain==0.0.248
pydantic==1.10.8
pyodbc==4.0.39